third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-htmlimports.html

<!DOCTYPE html>
<html>
<head>
<link rel="import" href="resources/import-hello.html">
<!-- import-hello.html shouldn't be shown twice as it shares same resource as above -->
<link rel="import" href="resources/import-hello.html">
<script src="../inspector-test.js"></script>
<script src="../resources-test.js"></script>
<script src="resource-tree-test.js"></script>
<script>

function test()
{
    ApplicationTestRunner.runAfterResourcesAreFinished(["resource-tree-test.js"], step1);

    function step1()
    {
        ApplicationTestRunner.dumpResourceTreeEverything();
        TestRunner.completeTest();
    }
}

</script>

</head>

<body onload="runTest()">
<p>
Tests resource tree model for imports.
</p>

</body>
</html>
